frontend:Add tooltip to default assignee selection
