+++ frontend/src/languages/french.js
@@ -201,7 +201,8 @@ const fr = {
     workflowitem_assignee: "Default assignee",
     organization_info: "Organization",
     total_budget_info: "Total budget",
-    default_assignee_warning: "Default assignee cannot be changed once Subproject is created"
+    default_assignee_warning: "Default assignee cannot be changed once Subproject is created.",
+    default_assignee_warning2: "Default assignee will be assigned to all wrokflow items in subproject without an option to change it."
   },
 
   workflow: {
