+++ frontend/src/languages/portuguese.js
@@ -285,7 +285,8 @@ const pt = {
     workflowitem_assignee: "Cessionário padrão",
     organization_info: "Organization",
     total_budget_info: "Total budget",
-    default_assignee_warning: "Default assignee cannot be changed once Subproject is created"
+    default_assignee_warning: "Default assignee cannot be changed once Subproject is created.",
+    default_assignee_warning2: "Default assignee will be assigned to all wrokflow items in subproject without an option to change it."
   },
 
   workflow: {
